Remove the "Player" tag from the BestFriend object.
  * This is what is causing all the things to not trigger properly since we only ever pick the first "Player" GameObject and the BFF loads first in the player.

Adjusted ProjectSettings for the player to use 1440x900 windowed mode by default, with option to switch to FullScreen mode.

Added some IEnumerator WaitForPlayerToSetup() coroutines in the SceneManagement and BestFriendTask as part of debugging (probably not needed but doesn't hurt).

Fixed the bug in LateUpdate() to also check for _healthHandler is not null before trying to get the component.